+++ services/libs/data-access-layer/src/organizations/identities.ts
@@ -2,6 +2,7 @@ import { QueryExecutor } from '../queryExecutor'
 import { QueryOptions, QueryResult, queryTable } from '../utils'
 import { getOrgIdentities } from './organizations'
 import { IDbOrgIdentity, IDbOrgIdentityInsertInput, IDbOrgIdentityUpdateInput } from './types'
+import { OrganizationIdentityType } from '@crowd/types'
 
 export async function fetchOrgIdentities(
   qx: QueryExecutor,
@@ -145,6 +146,33 @@ export async function upsertOrgIdentities(
       })
     }
   }
+}
+
+export async function findOrgIdByDomain(
+  qx: QueryExecutor,
+  tenantId: string,
+  domains: string[],
+): Promise<string | null> {
+  const result = await qx.selectOneOrNone(
+    `
+      SELECT "organizationId"
+      FROM "organizationIdentities"
+      WHERE "value" = ANY($(domains))
+        AND "tenantId" = $(tenantId)
+        AND "type" IN ('${OrganizationIdentityType.PRIMARY_DOMAIN}', '${OrganizationIdentityType.ALTERNATIVE_DOMAIN}')
+      LIMIT 1;
+    `,
+    {
+      domains,
+      tenantId,
+    },
+  )
+
+  if (result) {
+    return result.id
+  }
+
+  return null
 }
 
 export enum OrgIdentityField {
